+++ R/predict_general_mdl_avg_trend.R
@@ -54,6 +54,8 @@ predict_general_mdl_avg_trend <- function(df,
                                           sort_col = "year",
                                           sort_descending = FALSE,
                                           pred_col = "pred",
+                                          pred_upper_col = "pred_upper",
+                                          pred_lower_col = "pred_lower",
                                           upper_col = "upper",
                                           lower_col = "lower",
                                           filter_na = c("predictors", "response", "all", "none"),
@@ -92,6 +94,8 @@ predict_general_mdl_avg_trend <- function(df,
   assert_string(pred_col, 1)
   assert_string(upper_col, 1)
   assert_string(lower_col, 1)
+  assert_string(pred_upper_col, 1)
+  assert_string(pred_lower_col, 1)
   filter_na <- rlang::arg_match(filter_na)
   assert_string(types, 3)
   assert_string(source, 1)
@@ -121,8 +125,8 @@ predict_general_mdl_avg_trend <- function(df,
                                       sort_col = sort_col,
                                       sort_descending = sort_descending,
                                       pred_col = pred_col,
-                                      upper_col = upper_col,
-                                      lower_col = lower_col,
+                                      pred_upper_col = pred_upper_col,
+                                      pred_lower_col = pred_lower_col,
                                       filter_na = filter_na,
                                       ret = ret,
                                       error_correct = error_correct,
@@ -145,8 +149,8 @@ predict_general_mdl_avg_trend <- function(df,
                          obs_filter = obs_filter,
                          sort_col = sort_col,
                          pred_col = pred_col,
-                         upper_col = upper_col,
-                         lower_col = lower_col,
+                         pred_upper_col = pred_upper_col,
+                         pred_lower_col = pred_lower_col,
                          test_col = test_col)
 
   # Untransform variables
@@ -154,8 +158,8 @@ predict_general_mdl_avg_trend <- function(df,
     df <- probit_transform(df,
                            c(formula_vars[1],
                              pred_col,
-                             upper_col,
-                             lower_col),
+                             pred_upper_col,
+                             pred_lower_col),
                            inverse = TRUE)
   }
 
@@ -164,8 +168,8 @@ predict_general_mdl_avg_trend <- function(df,
     df <- scale_transform(df,
                           c(formula_vars[1],
                             pred_col,
-                            upper_col,
-                            lower_col),
+                            pred_upper_col,
+                            pred_lower_col),
                           scale = scale,
                           divide = FALSE)
   }
@@ -181,8 +185,8 @@ predict_general_mdl_avg_trend <- function(df,
                        sort_col = sort_col,
                        sort_descending = sort_descending,
                        pred_col = pred_col,
-                       upper_col = upper_col,
-                       lower_col = lower_col)
+                       pred_upper_col = pred_upper_col,
+                       pred_lower_col = pred_lower_col)
 
     if (ret == "error") {
       return(err)
@@ -197,6 +201,10 @@ predict_general_mdl_avg_trend <- function(df,
                          sort_col = sort_col,
                          sort_descending = sort_descending,
                          pred_col = pred_col,
+                         pred_upper_col = pred_upper_col,
+                         pred_lower_col = pred_lower_col,
+                         upper_col = upper_col,
+                         lower_col = lower_col,
                          type_col = type_col,
                          types = types,
                          source_col = source_col,
@@ -241,8 +249,8 @@ fit_general_average_model <- function(df,
                                       sort_col,
                                       sort_descending,
                                       pred_col,
-                                      upper_col,
-                                      lower_col,
+                                      pred_upper_col,
+                                      pred_lower_col,
                                       filter_na,
                                       ret,
                                       error_correct,
@@ -290,8 +298,8 @@ fit_general_average_model <- function(df,
       predict_general_data(df = y,
                            model = mdl,
                            pred_col = pred_col,
-                           upper_col = upper_col,
-                           lower_col = lower_col)
+                           pred_upper_col = pred_upper_col,
+                           pred_lower_col = pred_lower_col)
       })
 
     mdl <- NULL # not returning all models together for grouped models
@@ -307,8 +315,8 @@ fit_general_average_model <- function(df,
       df <- predict_general_data(df = grp_data,
                                  model = mdl,
                                  pred_col = pred_col,
-                                 upper_col = upper_col,
-                                 lower_col = lower_col)
+                                 pred_upper_col = pred_upper_col,
+                                 pred_lower_col = pred_lower_col)
     }
   }
   if (ret != "mdl") {
@@ -318,8 +326,8 @@ fit_general_average_model <- function(df,
                            sort_col = sort_col,
                            sort_descending = sort_descending,
                            pred_col = pred_col,
-                           upper_col = upper_col,
-                           lower_col = lower_col,
+                           pred_upper_col = pred_upper_col,
+                           pred_lower_col = pred_lower_col,
                            test_col = NULL,
                            error_correct = error_correct,
                            error_correct_cols = error_correct_cols,