+++ R/predict_simple.R
@@ -127,8 +127,8 @@ predict_simple <- function(df,
                        sort_col = sort_col,
                        sort_descending = sort_descending,
                        pred_col = pred_col,
-                       upper_col = NULL,
-                       lower_col = NULL)
+                       pred_upper_col = NULL,
+                       pred_lower_col = NULL)
 
     if (ret == "error") {
       return(err)
@@ -143,6 +143,10 @@ predict_simple <- function(df,
                          sort_col = sort_col,
                          sort_descending = sort_descending,
                          pred_col = pred_col,
+                         pred_upper_col = NULL,
+                         pred_lower_col = NULL,
+                         upper_col = NULL,
+                         lower_col = NULL,
                          type_col = type_col,
                          types = types,
                          source_col = source_col,